requirements.txt
certifi==2020.11.8
chardet==3.0.4
elasticsearch==7.10.0
flake8==3.8.4
idna==2.10
mccabe==0.6.1
oauthlib==3.1.0
py4j==0.10.9
pycodestyle==2.6.0
pyflakes==2.2.0
PySocks==1.7.1
pyspark==3.0.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
tweet-preprocessor==0.6.0
urllib3==1.25.11
vaderSentiment==3.3.2
